refactor(ensapi): extract validated primary-list resolution to a shared helper
Move the `primary-list` decode and two-step user-role validation out of the root
`efp.primaryList` resolver into `resolveValidatedPrimaryListTokenId`, so `Account.efp.primaryList`
(next commit) reuses the same logic. No behavior change.
